Add terminate_service function, and SpawnedFuture to support it
To asynchronously shut down services, we spawn a separate thread whose
sole job is to kill the service process and wait until it is
dead. Once it is dead, we use a oneshot channel to signal back that
the service has been shut down, and thus the parent future may
proceed.

A small `SpawnedFuture` abstraction is also added, which allows us to
model both the successful termination of the terminator thread, as
well as the failure to actually spawn that thread.

Signed-off-by: Christopher Maier <cmaier@chef.io>